requirements.txt
Django==5.1.5
djangorestframework==3.16.1
Pillow==10.4.0
requests==2.32.3
gunicorn